init.php modified, dynamic item interface, SharePreferences, Customer's Order module done.

+++ app/src/main/java/wqyap762/rprqs/OrderRequest.java
@@ -14,7 +14,7 @@ public class OrderRequest extends StringRequest{
     private static final String ORDER_REQUEST_URL = "http://rprqs.16mb.com/OrderFood2.php";
     private Map<String, String> params;
 
-    public OrderRequest(int total_price, int quantity, String payment_status, String username, String menu_id, Response.Listener<String> listener) {
+    public OrderRequest(Double total_price, int quantity, String payment_status, String username, String menu_id, Response.Listener<String> listener) {
         super(Request.Method.POST, ORDER_REQUEST_URL, listener, null);
         params = new HashMap<>();
         params.put("total_price", String.valueOf(total_price));
